rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG system for beach jurisprudence"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3",
    "langchain-google-genai>=3.0",
    "langchain-chroma>=1.0",
    "langchain-ollama>=0.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "chromadb>=1.0",
    "rank-bm25>=0.2",
    "google-genai>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[tool.uv]
package = false
